refactor: migrate to standalone components #3
