+++ src/main.ts
@@ -1,13 +1,58 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { enableProdMode, importProvidersFrom, Provider } from '@angular/core';
 
-import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
+import { AppComponent } from './app/app.component';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatBadgeModule } from '@angular/material/badge';
+import { CartModule } from './app/cart/cart.module';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { provideAnimations } from '@angular/platform-browser/animations';
+import { AppRoutingModule } from './app/app-routing.module';
+import { bootstrapApplication, BrowserModule } from '@angular/platform-browser';
+import { CONFIG_TOKEN } from './app/core/injection-tokens/config.token';
+import { ErrorPrintInterceptor } from './app/core/interceptors/error-print.interceptor';
+import {
+  HTTP_INTERCEPTORS,
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
+
+const interceptors: Provider[] = [
+  {
+    provide: HTTP_INTERCEPTORS,
+    useClass: ErrorPrintInterceptor,
+    multi: true,
+  },
+];
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
-  .catch((err) => console.error(err));
+bootstrapApplication(AppComponent, {
+  providers: [
+    importProvidersFrom(
+      BrowserModule,
+      AppRoutingModule,
+      MatToolbarModule,
+      MatButtonModule,
+      MatIconModule,
+      MatMenuModule,
+      MatTooltipModule,
+      CartModule,
+      MatBadgeModule,
+      MatSnackBarModule,
+    ),
+    interceptors,
+    {
+      provide: CONFIG_TOKEN,
+      useValue: environment,
+    },
+    provideAnimations(),
+    provideHttpClient(withInterceptorsFromDi()),
+  ],
+}).catch((err) => console.error(err));
